Create 330. Patching Array
Given a sorted integer array nums and an integer n, add/patch elements to the array such that any number in the range [1, n] inclusive can be formed by the sum of some elements in the array.

Return the minimum number of patches required.

Solution Explanantion:

Problem Explanation
Given an array nums of positive integers sorted in ascending order and a positive integer n, the goal is to determine the minimum number of patches (additional numbers) required so that every number in the range [1, n] can be represented as the sum of some elements from the array (including the patches).

Code Explanation
The function minPatches calculates the minimum number of patches needed. Here's a detailed explanation:

Variables
ans: This variable keeps track of the number of patches added.
i: This variable serves as the index for traversing the nums array.
miss: This variable represents the smallest sum in the range [1, n] that cannot be formed with the current elements of nums and the patches added so far.
Algorithm
The algorithm uses a greedy approach to determine the patches. It iterates while miss is less than or equal to n, indicating that there are still sums in the range [1, n] that cannot be formed.

While Loop:

The loop continues as long as miss is less than or equal to n.
Condition Check:

If the current element of nums (nums[i]) is less than or equal to miss, it means that nums[i] can help cover the current gap (miss). We add nums[i] to miss and move to the next element by incrementing i.
Otherwise, if nums[i] is greater than miss or if we've exhausted the array, we need to add a patch. The optimal patch to add is miss itself. This is because adding miss will extend the range of numbers that can be formed from [1, miss) to [1, 2*miss). We then increment ans (number of patches) and double miss.
Return Value:

After the loop completes, ans will contain the minimum number of patches required to ensure all numbers from 1 to n can be formed.
Concepts Involved
Greedy Algorithm:

The algorithm uses a greedy approach by always choosing the smallest missing sum (miss) as the next patch. This ensures the most significant increase in the range of representable numbers with each patch.
Dynamic Range Expansion:

By adding miss, the algorithm ensures that the range of numbers that can be represented is expanded optimally. For example, if miss is 4, adding it means all numbers up to 7 can be represented if all numbers up to 3 were previously representable.
Time Complexity:

The algorithm efficiently traverses the nums array and makes optimal additions in linear time relative to the size of nums.