Create 1482. Minimum Number of Days to Make m Bouquets
You are given an integer array bloomDay, an integer m and an integer k.

You want to make m bouquets. To make a bouquet, you need to use k adjacent flowers from the garden.

The garden consists of n flowers, the ith flower will bloom in the bloomDay[i] and then can be used in exactly one bouquet.

Return the minimum number of days you need to wait to be able to make m bouquets from the garden. If it is impossible to make m bouquets return -1.
Solution Explanantion:

Concepts and Explanation
Problem Understanding:

Inputs:
bloomDay: An array where each element represents the day a specific flower will bloom.
bouquetCount: The number of bouquets that need to be made.
flowersPerBouquet: The number of adjacent flowers required to make one bouquet.
Output:
The minimum number of days required to make the specified number of bouquets. If it is not possible to make that many bouquets, return -1.
Constraints:

If the total number of flowers (bloomDay.length) is less than bouquetCount * flowersPerBouquet, it's impossible to make the required bouquets. In such cases, the function returns -1.
Binary Search:

The problem can be solved efficiently using binary search over the range of days.
The range (minDays to maxDays) is determined by the minimum and maximum values in the bloomDay array.
Binary search helps to minimize the number of days needed by checking the mid-point (midDays) and adjusting the search range based on the feasibility of making the required number of bouquets within midDays.
Helper Method (calculateBouquets):

This method calculates how many bouquets can be made if flowers that bloom within waitDays are considered.
It iterates through bloomDay and counts the number of valid bouquets by checking consecutive flowers that bloom within waitDays.
If a flower blooms after waitDays, the count of required adjacent flowers is reset.
Detailed Steps:
Initial Checks:

If the total number of flowers is less than the required number of flowers for all bouquets, return -1.
Binary Search Setup:

Set minDays to the minimum value in bloomDay.
Set maxDays to the maximum value in bloomDay.
Binary Search Execution:

Calculate midDays as the average of minDays and maxDays.
Use calculateBouquets to determine if the required number of bouquets can be made within midDays.
Adjust the search range based on whether the required bouquets can be made:
If calculateBouquets returns a number greater than or equal to bouquetCount, search in the lower half by setting maxDays to midDays.
Otherwise, search in the upper half by setting minDays to midDays + 1.
Return the Result:

After exiting the loop, minDays will be the minimum number of days required to make the required bouquets.