Create 1791. Find Center of Star Graph
Solution Explanantion

The code defines a method findCenter that determines the center of a star graph from its edge list. In a star graph, one central node is connected to all other nodes. The method identifies this central node by examining the first two edges.

Explanation
Class and Method Definitions:

The Solution class contains the method findCenter.
findCenter(int[][] connections): This method takes a 2D array connections representing the edges of the graph.
Input:

connections is a 2D array where each sub-array represents an edge connecting two nodes in the format [u, v].
Logic:

The center of a star graph is the node that appears in all the edges.
By examining the first two edges (connections[0] and connections[1]), we can determine the center node:
If the first node of the first edge (connections[0][0]) is equal to either of the nodes in the second edge (connections[1][0] or connections[1][1]), then connections[0][0] is the center.
Otherwise, connections[0][1] is the center.
Code Explanation:

connections[0][0] == connections[1][0] || connections[0][0] == connections[1][1]: This condition checks if the first node of the first edge is present in the second edge.
? connections[0][0] : connections[0][1]: If the condition is true, the center is connections[0][0]. If not, the center is connections[0][1].
Example
Consider the input connections = [[1, 2], [2, 3], [4, 2]]:

The method will check the first two edges: connections[0] = [1, 2] and connections[1] = [2, 3].
It checks if 1 (first node of the first edge) is equal to 2 or 3 (nodes in the second edge). This is false.
It then determines that the center is 2 (the second node of the first edge) because 2 appears in both edges.